src/AccountingAgent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
openai>=1.60
anthropic>=0.42
PyMuPDF>=1.25
Pillow>=11.0
python-multipart>=0.0.18
python-dotenv>=1.0
